Correct build error with CGALlab
